Make forms novalidate by default
Built-in HTML5 form validations causes accessibility problems for some
users[0] and often are less-clear duplicates of the errors that Rails
would otherwise generate.

This commit flips the default to 'off' in builder-aware form methods
`form_for` and `form_with`. If for some reason it's necessary to enable
it, it can be overridden by passing `novalidate: false` into the HTML
options. The specs provide examples

[0] https://github.com/alphagov/govuk-design-system/issues/941