Standardise label weights and sizes
Instead of manually applying size and weight classes to labels we're now
using the GOV.UK Design System built-in label styles[0].

Weight is automatically applied to labels with these classes so the
option to configure it manually has been removed.

This work should also pave the way for building the labels/legends as
page headings[1] functionality.

[0] https://design-system.service.gov.uk/components/text-input/#label-text-inputs
[1] https://design-system.service.gov.uk/get-started/labels-legends-headings/