Display one error message per field in error summary
The error summary displayed all field errors grouped on one line.
This commit changes the error summary to only display the first
error for a field.

Edit the test Person object in the error_summary spec to be more concise

The error_summary spec uses a Person object to test validations. This commit changes how the Person object is created in the error_summary spec to be more concise.