Begin work on the formbuilder guide
While the docs are useful something a little more-digestable is probably
going to be more use to most people, it should give a taste of the code
required to produce various form components.

Keeping everything in sync is a pain, so this (experimental) approach is
intended to actually use the form builder itself to generate examples.
The output is a static site, generated by Nanoc[0] that 'loads' the form
builder and, for a supplied snippet, displays:
* the raw (input) code
* the rendered form group
* the compiled HTML

Currently, for simplicity's sake, everything is static. NPM isn't used
and the minified assets have been commtted. Perhaps this can be improved
on later.

[0] https://github.com/nanoc/nanoc

Refs #2