Allow labels to be optionally hidden
Sometimes labels add noise to a form and are unecessary. An example of
this is when capturing an address with multiple lines. The initial label
'Address' should make it obvious what lines 2-4 are for. The label
should still be present to keep the form accessible.

This change adds the ability to wrap the contents of a label in a span
that hides it visually. There is an example of this technique used in
the fieldset section of the Design System documentation[0]

[0] https://design-system.service.gov.uk/components/fieldset/