Add smallest_segment arg to govuk_date_field
The smallest_segment param refers to the smallest part of the date that
is represented by the group of inputs. It defaults to `:day` but that
can be overriden by `:month`. When overriden, the 'Day' part of the date
input is omitted and only the 'Month' and 'Year' segments are generated.

The ability to limit to just 'Year' hasn't been added, that is probably
better represented by a `govuk_number_field`.

Refs #56