Add support for flexible localisation schemas
The configuration block now supports the following keys for customising
how the builder's localisation functionality works:

* localisation_schema_fallback
* localisation_schema_legend
* localisation_schema_hint
* localisation_schema_label

The localisation_schema_fallback setting is the only one set by default,
and it will be used for legends, hints and labels if they are not
individually configured by their key respectively.

Configuration must be provided as an array; it will be suffixed by the
object name and attribute name.

The special value `__context__` will be substituted for the contexual
object; either label, hint or legend. It can be placed anywhere in the
schema.

Refs #80