Ensure errors when summary used on ruby objects
This will highlight that there's a problem if a developer attempts to
use the #govuk_error_summary in a form where the object doesn't respond
to #errors. Also reformat and tidy up the examples file.

The Naming/MethodParameterName cop is now disabled because it thinks two
letter method params are bad (we're using :cv) 🤦