Improve localisation
Extend the localisation schema with the `value` of the radio/checkbox option (for example "yes" or "no") to be able to localise these automatically.

This change should make it now possible to set separately labels and/or hints for either the fieldset or the individual radios/checkboxes inside that fieldset.

In order to not break the current interface, the `#govuk_collection_radio_buttons` method does not change, which means to be able to take the label from the locales, instead of providing to the method a `text_method`, just pass `nil`.

Example:

```ruby
f.govuk_collection_radio_buttons :parent, @values, :value, nil, inline: true
```

This makes for a bit ugly interface but could be improved later on. I didn't want to break compatibility.

Also, now localisation can also contain html elements, and in order to follow the usual Rails conventions, these locale keys will contain the suffix `_html` which will make the content html safe automatically when using `t()` method in the view, although it does not happen automatically outside the view that's why I use `#html_safe` if a locale was found.

Still, the non-html key will be tried first, and returned if one is found, otherwise the html version will be tried, and if none was found, nil will be returned as before.

Additionally I've done a small refactor so the collection check boxes hint elements are the same as with other form elements, by using the same class and method signature.

I couldn't find a reason why this was not the case, probably it was a leftover from a previous refactor.