Extend legend options to enable visually hidden
Allow to pass an optional option to the existing legend options hash, in a similar way to the one in labels, to add an extra class `govuk-visually-hidden` to the legend for those cases where we want to avoid visual repetition.

This is in line with recommendations from the design system:
https://design-system.service.gov.uk/get-started/labels-legends-headings/

However, this option only removes visual duplication and will not help users of screen readers. They will still hear both the page heading and the visually hidden `<label>` or `<legend>`.

To prevent this, set the contents of the `<label>` or `<legend>` as the page heading.