Fix bug where translations missing errors shown
When Rails is configured with `config.i18n.fallbacks = true`[0] the
behaviour of `I18n.exists?` is unclear[1]. It returns true when passed
`nil` because fallbacks are considered.

This could be fixed by specifying `fallback: false` as a keyword arg to
`I18n.exists?`[2] but if we have no key there's no point in proceeding to
the check stage.

[0] https://guides.rubyonrails.org/v6.0.0/configuring.html#configuring-i18n
[1] https://github.com/ruby-i18n/i18n/issues/365
[2] https://github.com/ruby-i18n/i18n/pull/482