Replace hardcoded 'govuk' with #brand
This is the first major step in making the form builder more flexible
with regards to the underlying design system. All hardcoded references
to `govuk` have been replaced with calls to the brand method, `prefix`
is used to simplify the branding of arrays of classes.

Manually checked that nothing in ./lib matches the following search:

$ ag -s govuk lib/govuk_design_system_formbuilder/{elements,containers,traits}

No new tests yet but existing ones pass!