Add support for nanoc live to ease guide writing
Nanoc has supported 'live editing' for a while but it hadn't been added
to the guide because I couldn't work out how to specify groups in the
gemspec (it needs to be in the `nanoc` group).

It turns out that adding it to the Gemfile works fine. The guide can now
be run using the make command `make watch-guide` or changing to the
`guide` directory and running `bundle exec nanoc live`.

This is a cleaner solution than the previous one of running two make
commands; one to serve the app and the other to rebuild on file changes.