Remove guide page for rebranding the form builder
It looks like the NHS design system isn't as close a match as initially
thought. The branding functionality is still a move in the right
direction but claiming support for the NHS DS is a step too far; an
wrapper library that amends the relevant HTML attributes will be
required in between.