Allow labels to be set via procs
Internally, the label class now takes an argument called content, this
overrides the text and other visual settings.

It should only set via label_args, it's not intended to be used on _all_
labels, like those used for checkboxes and radio buttons.

To maintain compatibilty with the design system's error handling, the
content of the proc will be wrapped in a <label> tag; this needs to be
the case because not allowing the form builder to be in control of
setting the 'for' attribute would lead to breakages - keeping the labels
associated with inputs would be left to the developer and this will lead
to inconsistency, complication and errors.

Refs #133