Add native captions to legend and label
This approach allows for an extra `caption` key in the hashes passed to
legend and label. If the caption hash has a value for the `text` key, a
caption will be rendered immediately before the label text, inside the
legend or label's tag - this appears to be the most-common use case for
captions.

https://design-system.service.gov.uk/styles/typography/#headings-with-captions

Refs #133