Make caption a top level keyword argument
Instead of passing caption in as part of the label or legend, move it up
a level. This approach has a number of advantages including allowing for
more flexbility with captioning labels that are generated from sources
other than the passed-in args (ie localised or defaults) and it
simplifies the documentation.

Now, captions can be added in the following manner:

= govuk_text_field :name, caption: { text: 'Stage one' }

Refs #133, #135