Simplify HTML building
This is an attempt to make the construction of the HTML output more
obvious. The previous approach worked well for smaller, simpler inputs
but some of the more-complex ones were difficult to read. This was made
more difficult by the multiple layers of indentation for hashes and
longer arrays.

These have now, for the most part, been separated out to their own
methods. The naming convention and structure is now much simpler and
easier to follow.

For consistency, _args methods have been renamed to _options (as they're
the options that configure the element) and what were formerly
extra_args are now kwargs.