Allow to set a default caption size via configuration
This follows the same pattern used for legends.

It is possible to set a default, project-wise, caption size by using the configuration block.

I've maintained the default as `m` (as that was the chosen size before, and also the default for legends) in order to not break any services that might be already using this size.

In our service the default for captions, as these are used in all cases with h1 page headers, will be `xl` and this PR will save us from having to manually specify the size in each page we use captions.

Note: I will be sending a separate PR in a few minutes for a separate issue we encounter with localisation of captions. But wanted to split the work for easier review.