Fix caption leaking its localisation to radio options
We run into a problem when using the new captions capability of the gem [1] where the localised caption of `govuk_collection_radio_buttons` or `govuk_radio_buttons_fieldset` will leak also to each of their corresponding radio option labels.

I think there is not a use-case to use captions in the labels of individual radio or check box options and couldn't find anything in the design system documentation about this, so I opted for the simplest possible solution, but if in the future there is a use-case for having captions also in these option labels, then I suppose a bit of work is neccessary in the localisation to not inherit that of the fieldset.

I didn't see this happening to check boxes, at least in the scenarios I have, but just in case I'm also excluding these check boxes from being "captionable".

[1] https://github.com/DFE-Digital/govuk_design_system_formbuilder/pull/135