Add support for customisable form group attributes
Instead of only allowing a form group's classes to be configured, this
change replaces the form_group_classes argument with form_group which
takes a hash.

Adding a classes key to the hash will append these extra classes to the
required ones (.govuk-form-group). Any other key/value pairs will be
added as attributes to the form group element.