Wrap the calling of blocks in capture
This bug was introduced in 08fe4ffc05 and wasn't picked up in the tests,
probably due to the fact we're not using a templating layer here just
plain Ruby. It turns out capture is needed to prevent the block content
from appearing once where it's meant to and sometimes immediately when
the block is called.