Display ERB examples alongside Slim ones
This was brought up in another issue and there was a suggestion that
Slim templates might not make sense to people using ERB. After a bit of
exerimentation it seems that with a bit of _manipulation_  the output is
useable.

The main thing that could be improved is the indentation of arguments
split across several lines. For the moment I've put them all on one line
because the HtmlBeautifier library isn't indenting them correctly. Most
examples are still readable.

Refs #181