Set default legend text to attribute unless nil
To make the behaviour of fieldset legends more consistent with labels,
they will now be added to the form by default if nothing is specified.

This is intended to act as a prompt for the developer to modify the
content, so no additional formatting (ie removing underscores) is
undertaken.

If the developer wants the legend to be omitted they can pass `nil` to
in legend param.

Fieldsets with no attribute name (ie standalone fieldsets called via
`govuk_fieldset`) will remain unaffected.

Fixes #142