Move instance variable assigns out of else block
Some of the instance variables set in Label's initializer are required
for both proc and hash style labels. When set using a proc they are
missed and it causes the ids and radio/checkbox-specific classes not to
be set properly.

They have now been moved out, only the variables used to configure the
appearance of the label are now set in the else clause.