Fix localisation of collection labels/hints
This fixes a bug where @value wasn't being assigned when the hint_method
param wasn't passed in, even though @value is required to localise it
properly.

Tests have been added to ensure it works as expected; for radio
collections both the label and hint can be set via localisation and for
check boxes just the hint can.