Fix bug where checkbox hidden field isn't rendered
When web forms are submitted, check boxes in their unchecked state are
not submitted[0] so Rails' `#check_box` helper, which we wrap in
`#govuk_check_box`, should automatically create a hidden field with the
unchecked value.

This has not been happening because Rails omits the hidden field when
the unchecked value is `false`. This wasn't covered by the test suite.
The documentation recommends that a value of `0` should be used to
represent it instead[1]

This fix does two things:
* changes the default unchecked value from `false` to `0`
* allows custom unchecked values to be set in a way that matches Rails'
  API

[0] https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/checkbox#Value
[1] https://api.rubyonrails.org/classes/ActionView/Helpers/FormBuilder.html#method-i-check_box