Add missing hidden field needed for deselection
When forms are submitted, when a field has no check boxes checked it's
omitted from the form's payload[0]. To combat this, Rails adds a hidden
field with no value before the check boxes are rendered. This change
makes the form builder do the same and adds a check to ensure its
presence.

This behaviour is already provided by #govuk_collection_check_boxes
which wraps Rails' functionality more closely.

[0] https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/checkbox#Value