Render button when HTML content provided to submit
The `#govuk_submit` method usually renders an `<input type='submit'>`.

Sometimes it might be useful to allow HTML to be included inside a
form submission button which would be supported by a `<button>` tag
instead.

When the submit content is just text it seems that there's no advantage
in using a button over an input, so we'll stick to using what Rails
expects.

This change updates the behaviour so when `#govuk_submit` is supplied
with HTML (via a proc), it render the contents inside a button tag.

This deviates slightly from Rails' form builder which always uses an
input, and there *may* be issues with the way IE6 deals with buttons,
but it shouldn't cause any major problems with modern browsers.

Refs #205