Add example of localised check boxes to guide
There was already an example but the existing one was simpler, using the `govuk_collection_check_boxes`.

This new example focus more in the `govuk_check_boxes_fieldset` declaring individual check boxes and one with a revealing text box. Each of the check boxes have independent labels and hints localised.