Allow suppression of elements by passing in nil
This addresses a problem where elements are always rendered if there's a
matching localisation key, even if `nil` is supplied to the helper.

Introducing a Elements::Null class simplifies the logic - an alternative
solution was to introduce an :enabled flag but as we already have
:active it was noisy and confusing.

When the relevant argument (:caption, :hint, :label or :legend) is nil a
Elements::Null is built instead of the usual 'active' element. It will
never render any HTML and allows the suppression of elements.

The default argument for captions, hints, labels and legends is always
an empty hash, so simply not providing any arguments will allow the
localisation mechanism to kick in and populate the element.

Fixes #212