Make the specs work for old and new Rails errors
Rails 6.1.0 introduced a new structure for error messages[0] that
substantially changed the ways errors are accessed and manipulated[1].
The library only retrieves errors via `object.errors.messages` which
didn't really change too much, but the specs did; this commit addresses
those changes.

As we'll need to support both older (5.2.x) versions of Rails for the
foreseeable future, multiple ways of adding and removing messages need
to be implemented. This is governed by the utility method
`rails_version_later_than_6_1_0?` which checks the regular
`RAILS_VERSION` environment variable approach but should default to the
latest version.

To ease development and testing, `.envrc` is now ignored so direnv[2,3]
can be configured as required.

[0] https://github.com/rails/rails/pull/32313
[1] https://code.lulalala.com/2020/0531-1013.html
[2] https://direnv.net/
[3] https://github.com/direnv/direnv