Upgrade to SimpleCov from 0.17.1 to 0.20.0
This was a non-trivial upgrade because versions 0.18 and 0.19 of
SimpleCov aren't compatible with CodeClimate, but <= 0.17.1, and
>= 0.20.0 are.

To make SimpleCov output a 'JSON Coverage report' (in coverage/coverage.json)
the CC_TEST_REPORTER_ID needs to be present in the RSpec workflow step.

I'm not 100% sure whether it's required for before-build, but added it
anyway for good measure.