Fix collection helper handling of symbol values
When a value is passed into the collection helper via a proc as a
symbol, it wasn't applied properly when constructing the elements
within. This could lead to inputs not having a value, labels not linking
to the corresponding input and hints not describing their intended
target.

This change ensures that any symbols are converted to strings and adds
some tests ensuring that when symbols are supplied via procs the output
is rendered properly.