Wrap submit and additional buttons in group
Version 3.11.0 of the GOV.UK Frontend introduced button groups[0]. The
grouping makes it easier to lay out buttons and links side-by-side than
previously, which required a margin to be added to the right of each
button that had another after it.

The ButtonGroup class is currently bare but it _may_ be useful if we
extend it to allow extra classes or attributes to be set on the group.

[0] https://github.com/alphagov/govuk-frontend/releases/tag/v3.11.0