Add the default_submit_validate config option
It will control whether or not the element generated by #govuk_submit
will have the formnovalidate attribute set or not. This is important
because many projects are using custom #form_for/#form_with setups to
enforce the novalidate attribute on their forms.

Having novalidate set on the form and formnovalidate set on the submit
input isn't what people would expect by default, so this gives
developers more of an option.