Drop Rails 5.2.4 from the supported versions
Rails 5.2.4 doesn't work with Ruby 3.0.0. Rails 5.2.0 was released
nearly three years ago and most projects using this library are on later
versions.

This is only dropping *official* support, it's likely that the gem will
continue to work for the time being but Rails 5.2.4 will no longer be
tested as part of the build.