Ensure classes aren't overwritten by kwargs
This commit introduces a small dependency on the deep_merge gem, which
when run via Rails compat mode gives us the Hash#deeper_merge. This has
the benefit of merging arrays within hashes, allowing us to combine the
default classes generated by the formbuilder with the additional ones
that might be passed in.