Make date field value setting more lenient
A recent change caused test failures in some projects because 'odd'
values were used to set dates. These included a Struct with day, month
and year accessors that allows nil values, as found in the Apply For
Postgraduate Teacher Training project.

Now we're calling the method directly if the attribute responds to the
segment; as this is now _within_ the if statement rather than a chain of
ors, if the value is nil, nil will be returned.

There's an additional failure state to capture and inform when objects
that are un-datelike are provided.

Fixes #259