Add option to allow date value to log when invalid
Currently it only fails. This is a sensible choice for most projects but
is likely to become a reason why those in the process of reworking their
codebase can't stay up to date with the gem.

The new option, enable_log_on_invalid_date, is disabled by default so
the behaviour of the library won't change uninvited.

Fixes #266