Ensure no conditional attrs added when no block
When either no block or an empty block (i.e., one that fails
Rails' #presence check) is provided in the block passed to a radio
button or check box no conditioal content shouldn be rendered and the
data-attribute that's used to toggle it should be omitted.

This tweak moves the logic to earlier in the object's initialisation so
that both the @conditional_content and @conditional_id will not be set
when the block is blank.

Fixes #281, 282