Add a helper that renders an error summary
This should be useful in instances where an error summary is needed
outside of a form. Currently the only way is to extend the form so it
contains the content above; this works and doesn't really make much
difference to users but it's a bit untidy.

We're creating a new class, FormBuilderProxy, to handle on-the-fly
rendering of objects that would usually reside in a form. It currently
just inherits FormBuilder, but we may want to customise or limit this
later - and the naming should make it clear from the helper's point of
view that it's transient and shouldn't really be used like a typical
form builder object.