Implement custom sorting of error messages
Currently this depends on there being a #error_order method being
present on the object/model associated with the builder. I went for this
approach so that with objects that are represented by several forms
through an app, there's no duplication in the views. It may be worth
reconsidering this, but it feels broadly sensible.

The way ordering works is the #error_order method should return an array
of attribute names, eg: [:name, :email, :age] and when building the list
of errors the error summary promotes these to the start.

Any errors on attributes that aren't covered by #error_order will appear
after those that are, in the order they were defined in the object or
model.

If there is no #error_order method on the object, don't try to re-order
and just go with what Rails gives us, which is the order they were
defined.