Improve the error sorting index behaviour
Although it's unlikely to happen, using the number of error messages as
the offset when placing unordered errors at the end of the list is the
wrong strategy. If there are enough attributes before the ones we care
about in the order arg, we can result in:

total errors + count < ordered error index

This leads to the ordered and unordered errors being shuffled together.

This is now avoided by creating an adjustment that's the combined total
of error messages and custom ordered attributes, and adding the error
counter to it.