Remove block capture from `radio_buttons_fieldset`
Inherits from base which captures the `block.call` already.

Because `super(arg1,arg2,etc)` is called, the block is passed implicitly
whether you supply it as an arg to super, `&block`, or not. This
results in the block being called and captured twice `capture { block.call }`,
once for the fieldset and once by superclass `Base`.

In the case of a block being passed that renders a partial it
means the partial will be rendered twice for example.

```
f.govuk_radio_buttons_fieldset do
  = render partial: 'my_radio_button_partial' => gets rendered twice, at least
```