Handle passing of &block consitently and raise errors when block required
check box and radio button fieldsets were handling passing of blocks
either explicitly (check_boxes_fieldset) or implicitly (radio_buttons_fieldset).

This handles blocks being passed to superclass implicitly for both for consitency.
 - could alternatively pass them explicitly?!

A couple of specs broke indicating current functionality expects an error to
be raised when no block is passed the helpers. I have handled this by raising
an error explcitly for these objects, and by raising a `LocalJumpError`.

`LocalJumpError`s are raised when yield is called in a method which is not supplied
with a block, so is, perhaps, a more idiomatic error?!

`block_given?` is used to test whether block is provided to avoid side affects
as it does not call the block, [I believe](https://ruby-doc.org/core-2.7.0/Kernel.html#method-i-block_given-3F)