Remove capture of blocks from `fieldset_item`
The `capture { block.call }` in the `conditional_content`
method causes blocks passed to `govuk_radio_buttons_fieldset`
or `govuk_check_boxes_fieldset` to be rendered twice.

This is because the `Traits::FieldsetItem` is included in
`FieldsetRadioButton` and `FieldsetCheckBox`, which inherit
from `Base` and call super(), which captures the block content
already.

```
@block_content = capture { block.call } if block_given?
```

This change assumes that `Traits::FieldsetItem` will only
ever be included in Objects that already capture block content
and can therefore pass their content directly.