Add a default error summary presenter
The intention of this presenter is that it provides the current basic
behaviour but also provides a convenient way for people with specific
requirements to inject them.

It currently requires that it's initialized with an error_messages hash
but we may yet tweak that so an instance can be passed in and the error
messages set on it.

Refs #310