Allow any object to be supplied as a presenter
This approach offers a little more flexibility when than just accepting
a class. When a class is provided it's instantiated with the object's
error messages but when an object is supplied it's used as-is.

It needs to implement the method #formatted_error_messages so we check
for that, but any other attributes and state can be set up in the
application and can contain whatever's necessary.