Make radio labels respect the bold_labels setting
The previous behaviour was added very early on in the library's
development and has remained unchanged since. It attempted to improve
the contrast between the labels and hints by, when both are present,
making the labels bold.

Unfortunately in doing so, it ignores the user preference when hints are
present and non-bold labels are wanted.

This change updates the `bold_labels` argument default from `false` to
`nil`. When left `nil`, the old behaviour is maintained, but now when
set to either `true`/`false` the selected behaviour will take
precedence.