Switch from list of unmergeable to mergeable attrs
Previously we held a list of HTML attributes that should never be merged
together by this library. It's clear that the opposite approach makes
more sense, there's only a small(ish) number of HTML attributes that we
should try - those that allow a space-separated list of items.

The common examples are class, aria-describedby and aria-labelledby, but
there are quite a few other aria ones[0].

Fixes #318

[0] https://www.w3.org/TR/wai-aria-1.0/states_and_properties#state_prop_def