Improve the tests around merging HTML attributes
This adds a few more tests that try to ensure we're properly dealing
with custom HTML attributes.

There are a couple of small improvements to area labelledby and
describedby argument building where we wrap the id in an array to allow
the values to be cleanly merged if required.

Also add a test example with some JSON for completeness.