Allow turbolinks data attribute to be customised
Previously this was hardcoded to 'data-turbolinks' but now Turbo[0] is
gaining traction it makes sense to add support. This approach will also
allow other 'pushstate'/pjax tools to be used if necessary.

The new config item default_error_summary_turbo_prefix supports being
set to either a string or nil.

When it's a string of 'turbo', every link in the error summary will have
the attribute `data-turbo="false"`. The default value is 'turbolinks'

When nil no data attribute will be added to the link.

[0] https://turbo.hotwired.dev/