Improve class building using a utility method
This improvement follows on from a recent PR on the components lib[0]
where Rails' class_names[1][2] was used to improve the building of lists
of HTML classes.

Unfortunately it returns a space-separated string instead of an array -
and the method it uses to actually build up the classes,
`#build_tag_values`, is private, which makes utilsing it a bit of a
pain.

Instead here we're implementing a minimal version of it called
`#build_classes`. It uses more-or-less the same class-building strategy
as `#build_tag_values` but is less flexible in terms of args it accepts.

[0] DFE-Digital/govuk-components/pull/264
[1] https://github.com/rails/rails/pull/37918
[2] https://www.bigbinary.com/blog/rails-6-1-introduces-class_names-helper