Allow error presenters to override link targets
Most of the time we want to link to an error that's on the same page
(and managed by the form builder itself) but there may be occasions
where we need to do something a little more specialised, such as linking
to something outside of the form builder or linking to another page
entirely.

The format for #formatted_error_messages is an array of arrays that
contain two items; the first is the attribute and the second is the
corresponding error message.

Here we're implementing a optional third item that when set will form
the error's link target. This will allow developers a greater level of
customisation while keeping the library generic.