Support the maxlength attribute when building the Date element
We want to be able to control the number of characters the user is able
to input for the day, month and year inputs.

This change enables that with an extra parameter to the govuk_date_field
method whilst defaulting to the old behaviour of not including dates.

It is true that the year input can be valid above four characters but
we're so far from that that it made sense to just force it at 4 chars
and making the choice for the user.