Allow error message trust to be toggled
Error messages are not trusted by default - they are not marked as
html_safe. This is because the user input might be included in the
error message.

Including the user's input in the message (e.g., "31 02 2019 is not a
valid date") is not a common practice in GOV.UK projects and sometimes
teams might want to add extra formatting, like a line break, to an error
message to make it easier to read.

To support this, this change introduces a toggle in the config. When
true, it will call `#html_safe` on all error messages before they're
injected into the output - this will maintain any HTML tags.

The default value is false - no behaviour will change unless the setting
is manually enabled.