Change postinstall cp from -r to -R
BSD's cp is archaic, POSIX[0] describes the behaviour as:

> If the -r option was specified, the behavior is implementation-defined

Here GNU cp has different behaviour from BSD cp (as found on macOS), the
former copies the source to the destination and the latter copies the
contents of the first to the second.

To fix it, both set the recursive flag with -R and add the asterisk to
the copy source so we're only going to pull the contents of the assets
dir and not the dir itself.

[0] https://pubs.opengroup.org/onlinepubs/009695399/utilities/cp.html