Add a sitemap to the guide using Nanoc's helper
The Nanoc helper makes this really easy, we're using Nanoc's list of
items so partials aren't included and we just need to filter to slim
files in `xml_sitemap`.

This doesn't set `changefreq` - not sure how important they are. If we
add them they needed to be added to each page's frontmatter
individually, probably set to `yearly`