Use html_attributes_utils gem over deep_merge
The deep_merge gem is incredibly powerful and versatile but isn't the
best fit for the specific task of merging HTML attributes while paying
attention to some special rules, like the splitting and merging of
certain attributes. This is useful in things like `class:`, where a list
of values is represented by a space-delimited string.

The new gem, `html_attributes_utils`, was created using the code removed
in this commit as a basis. It will benefit from being easier to test
standalone and can be reused in the govuk-components library too - and
maybe beyond!