Add config to control boldening of radio labels
The original behaviour of the form builder followed the examples in the
Design System documentation at the time and emboldened any radio button
labels when hints were present.

These examples were dropped in 2020[0] and the guidance changed. In the
referenced PR I said I'd look into changing the behaviour in the form
builder but never did.

As it would be a breaking change we're just adding a new configuration
option, `default_collection_radio_buttons_auto_bold_labels`, to allow
the behaviour to be modified.

[0] https://github.com/alphagov/govuk-design-system/pull/1284