Stop using GDS Transport and the crown logo
It was brought to my attention that the form builder guide doesn't
follow the rules for community resources and tools[0] with regards to
fonts and the use of the crown logo.

This puts the library in a slightly difficult situation because it's
raison d'être is to make building apps that follow the Design System
easier.

The crown logo has been removed and GOV.UK in the header replaced with
'Department for Education'.

On the typography front, the relevant section of the Design System
docs[1] states:

> If your service is publicly available on a subdomain other than
> service.gov.uk, use an alternative typeface like Helvetica or Arial.

Neither of the suggested fonts look close like GDS Transport. Not
looking like the thing we're building for would be a huge failure of the
guide; it might even make people think they're incompatible.

Instead my company (Graphia Ltd), has bought licences for New
Transport[2], the font GDS Transport is based on. While not exact, they
are similar enough that casual users probably won't notice a difference.

[0] https://design-system.service.gov.uk/community/contribution-criteria/#developing-a
-community-resource-or-tool
[1] https://design-system.service.gov.uk/styles/typography/
[2] http://www.newtransport.co.uk/