Add an inset paragraph to the guide on other tools
Some people inevitably come to the guide from Google thinking it's a
"form builder" in the non-developer sense (a tool people can use to
build forms).

The library uses the term in the Rails sense (as it wraps Rails'
FormBuilder).

Linking to consumers of this library is probably a good idea.