Add config flag to disable nested localisation
Scanning the object's name to work out whether or not it's nested should
work fine and be safe, but as this will affect every piece of text
rendered by the library there's a chance some edge cases might be
problematic.

As a result, add a config flag that disables the behaviour. If
everything goes smoothly this can be removed in an upcoming release.