Fix spacing on code examples in the guide
The `code` selector is adding some padding to the code blocks as well as
the inline code, which makes the first line slightly more indented than
those that follow. It's only really noticable in some examples (like the
error) where there are a few lines.

This change targets the code tags directly in paragraphs, leaving blocks
unaffected.