Add warning on Rails relationship presence checks
This problem has occurred since Rails added default validation to Rails'
belongs_to relationships in version 6.

Usually when creating forms we set relationships using the foreign key
column, i.e. `department_id`, not the `department` relationship
attribute itself .

This mismatch causes errors on the object to be reported in the error
summary but they won't link properly to the input because it's mapped to
a different field.

Solving the problem itself is complex so this is a tentative first step
and adds warnings to the guide.