Allow legend size to be nil
This updates the legend element to allow the size attribute to be `nil`,
and when `nil` a size class won't be specified for the element. This is
required to support the case where a field has a legend element but
shouldn't have a size class.

There's an example of this kind of component in the GOV.UK Design
System:

https://design-system.service.gov.uk/components/date-input/#if-youre-asking-more-than-one-question-on-the-page

It's currently not possible to re-create this example using purely the
form builder classes as the size is enforced as one of `s`, `m`, `l` or
`xl`.